+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * KLS-271 - Remove past media for completed export academy events
 * KLS-430 - EA Event details page
 * EAM-104 - Join and view buttons
+* KLS-388 - Added CTA streamfield block
 
 
 ## [2.28.0](https://github.com/uktrade/great-cms/releases/tag/2.28.0)